removed kw args from init methods for SingleCube, Ingredient, and Agent
