+++ cellpack/autopack/ingredient/single_cube.py
@@ -1,4 +1,4 @@
-from math import sqrt
+from math import sqrt, pi
 import numpy
 from panda3d.core import Point3, TransformState, Vec3
 from panda3d.bullet import BulletBoxShape, BulletRigidBodyNode
@@ -17,49 +17,93 @@ class SingleCubeIngr(Ingredient):
 
     def __init__(
         self,
-        molarity=0.0,
-        radii=None,
-        positions=[[[0, 0, 0], [0, 0, 0], [0, 0, 0]]],
-        positions2=[[[0, 0, 0]]],
-        sphereFile=None,
-        packingPriority=0,
-        name=None,
-        pdb=None,
-        color=None,
-        nbJitter=5,
-        jitterMax=(1, 1, 1),
-        perturbAxisAmplitude=0.1,
-        principalVector=(1, 0, 0),
-        meshFile=None,
-        packingMode="random",
-        placeType="jitter",
         Type="SingleCube",
-        meshObject=None,
+        color=None,
+        coordsystem='right',
+        cutoff_boundary=None,
+        cutoff_surface=None,
+        distExpression=None,
+        distFunction=None,
+        encapsulatingRadius=0,
+        excluded_partners_name=None,
+        force_random=False,  # avoid any binding
+        gradient="",
+        isAttractor=False,
+        jitterMax=(1, 1, 1),
+        # meshFile=None,
+        # meshName=None,
+        # meshObject=None,
+        # meshType='file',
+        molarity=0.0,
+        name=None,
+        nbJitter=5,
         nbMol=0,
-        **kw
+        offset=None,
+        orientBiasRotRangeMax=-pi,
+        orientBiasRotRangeMin=-pi,
+        overwrite_distFunc=True,  # overWrite
+        overwrite_nbMol_value=None,
+        packingMode="random",
+        packingPriority=0,
+        partners_name=None,
+        partners_position=None,
+        pdb=None,
+        perturbAxisAmplitude=0.1,
+        placeType="jitter",
+        positions2=[[[0, 0, 0]]],
+        positions=[[[0, 0, 0], [0, 0, 0], [0, 0, 0]]],
+        principalVector=(1, 0, 0),
+        proba_binding=0.5,
+        proba_not_binding=0.5,  # chance to actually not bind
+        properties=None,
+        radii=None,
+        rejectionThreshold=30,
+        resolution_dictionary=None,
+        rotAxis=None,
+        rotRange=6.2831,
+        source=None,
+        # sphereFile=None,
+        useOrientBias=False,
+        useRotAxis=False,
+        weight=0.2,  # use for affinity ie partner.weight
     ):
-        Ingredient.__init__(
-            self,
-            molarity=molarity,
-            radii=radii,
-            positions=positions,
-            positions2=positions2,
-            sphereFile=sphereFile,
-            packingPriority=packingPriority,
-            name=name,
-            pdb=pdb,
-            color=color,
-            nbJitter=nbJitter,
-            jitterMax=jitterMax,
-            perturbAxisAmplitude=perturbAxisAmplitude,
-            principalVector=principalVector,
-            meshFile=meshFile,
-            packingMode=packingMode,
-            placeType=placeType,
-            meshObject=meshObject,
-            nbMol=nbMol,
+        super().__init__(
             Type=Type,
-            **kw
+            color=color,
+            cutoff_boundary=cutoff_boundary,
+            cutoff_surface=cutoff_surface,
+            distExpression=distExpression,
+            distFunction=distFunction,
+            excluded_partners_name=excluded_partners_name,
+            force_random=force_random,
+            gradient=gradient,
+            isAttractor=isAttractor,
+            jitterMax=jitterMax,
+            # meshFile=meshFile,
+            # meshObject=meshObject,
+            molarity=molarity,
+            name=name,
+            nbJitter=nbJitter,
+            nbMol=nbMol,
+            overwrite_distFunc=overwrite_distFunc,
+            overwrite_nbMol_value=overwrite_nbMol_value,
+            packingMode=packingMode,
+            packingPriority=packingPriority,
+            partners_name=partners_name,
+            partners_position=partners_position,
+            pdb=pdb,
+            perturbAxisAmplitude=perturbAxisAmplitude,
+            placeType=placeType,
+            positions2=positions2,
+            positions=positions,
+            principalVector=principalVector,
+            proba_binding=proba_binding,
+            proba_not_binding=proba_not_binding,
+            properties=properties,
+            radii=radii,
+            rotRange=rotRange,
+            # sphereFile=sphereFile,
+            weight=weight,
         )
         print("MAKING CUBE", positions2)
         if name is None:
